requirements.txt
asgiref==3.5.2
attrs==22.2.0
Django==4.1.4
django-cors-headers==3.13.0
django-guardian==2.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
inflection==0.5.1
jsonschema==3.2.0
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2022.7
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
